root-platform/app/repositories/user_repository.py
from app.models.user import User
from app.extensions import db
from sqlalchemy import or_

class UserRepository:
    @staticmethod
    def get_all():
        return User.query.all()

    @staticmethod
    def get_by_id(user_id):
        return User.query.get(user_id)

    @staticmethod
    def get_by_username_or_email(identifier):
        return User.query.filter(
            or_(User.username == identifier, User.email == identifier)
        ).first()

    @staticmethod
    def create(user):
        db.session.add(user)
        db.session.commit()
        return user

    @staticmethod
    def update():
        db.session.commit()

    @staticmethod
    def delete(user):
        db.session.delete(user)
        db.session.commit()
