root-platform/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
requests
psycopg2-binary
aws-wsgi
Werkzeug
MarkupSafe
boto3
Flask-CORS
Flask-Injector
Flask-Cognito
flask_jwt_extended